+++ src/components/MapStyleInputWrapper.svelte
@@ -6,21 +6,27 @@
     config as configStore,
   } from '../stores';
   import { createBranchUrl } from '../branch-utils';
+  import { getRenderers } from '../renderers';
   import MapStyleInput from './MapStyleInput.svelte';
 
   export let index;
+  export let stylesheet;
 
   let branchPatterns;
   configStore.subscribe(value => ({ branchPatterns } = value));
 
   let map;
-  let url;
   let branch;
+  let renderer;
+  let type;
+  let url;
   let stylePresets = [];
   mapsStore.subscribe(maps => {
     map = maps.find(m => m.index === index);
     if (map) {
       branch = map.branch;
+      renderer = map.renderer;
+      type = map.type;
       url = map.url;
     }
   });
@@ -32,9 +38,27 @@
 
   let selectedValue;
 
+  let rendererOptions = [];
+  let rendererValue;
+
   // Selects the appropriate value from dropdownValues and adds any necessary key
   const setSelectedValue = () => {
-    let nextValue = dropdownValues.find(item => !!item.selected);
+    let nextValue;
+
+    for (const dropdownValue of dropdownValues) {
+      if (!!dropdownValue?.selected) {
+        nextValue = dropdownValue;
+      }
+      if (dropdownValue?.type === 'sublist') {
+        for (const dv of dropdownValue?.presets) {
+          if (!!dv?.selected) {
+            nextValue = dv;
+          }
+        }
+      }
+    }
+
+    if (!nextValue) return;
 
     // Don't mutate the dropdown values
     nextValue = JSON.parse(JSON.stringify(nextValue));
@@ -57,6 +81,30 @@
     selectedValue = nextValue;
   };
 
+  // Set renderer options based on the selected map's type
+  const setRendererOptions = () => {
+    rendererOptions = getRenderers(selectedValue, stylesheet?.sources);
+    setRendererValue();
+  };
+
+  // Set renderer value based on the selected map's renderer or type if no
+  // renderer is set
+  const setRendererValue = () => {
+    let updatedRenderer = renderer;
+
+    const validRenderers = getRenderers(selectedValue, stylesheet?.sources).map(
+      r => r.value
+    );
+
+    if (!updatedRenderer || !validRenderers.includes(updatedRenderer)) {
+      updatedRenderer = validRenderers.includes(selectedValue.type)
+        ? selectedValue.type
+        : validRenderers[0];
+    }
+
+    rendererValue = updatedRenderer;
+  };
+
   // Creates dropdown values and display options for the style dropdown linked by ids
   const setInitialDropdownOptions = () => {
     // This can be called multiple times on load, so always start fresh to prevent
@@ -69,21 +117,45 @@
       const stylePresetValues = stylePresets.map(item => ({
         ...item,
         dropdownType: 'preset',
-        selected: url === item?.url && map.type === item?.type,
+        selected: url === item?.url && type === item?.type,
         dropdownId: hat(),
+        ...(item.type === 'sublist' && {
+          presets: item.presets.map(v => ({
+            ...v,
+            dropdownType: 'preset',
+            selected: url === v?.url && type === v?.type,
+            dropdownId: hat(),
+          })),
+        }),
       }));
 
       dropdownValues = dropdownValues.concat(stylePresetValues);
 
-      dropdownDisplayOptions['Presets'] = stylePresetValues.map(item => ({
-        text: item.name,
-        dropdownId: item.dropdownId,
-      }));
+      dropdownDisplayOptions['Presets'] = stylePresetValues.map(item => {
+        return {
+          text: item.name,
+          ...(item.type !== 'sublist' && { dropdownId: item.dropdownId }),
+          ...(item.type === 'sublist' && {
+            type: 'sublist',
+            presets: item.presets.map(v => ({
+              text: v.name,
+              dropdownId: v.dropdownId,
+            })),
+          }),
+        };
+      });
     }
 
-    // Create values and displays for branch options
     if (branchPatterns) {
+      let patterns = [];
       for (const pattern of branchPatterns) {
+        let preset = {
+          name: pattern.name ?? pattern.id,
+          type: 'sublist',
+          dropdownId: hat(),
+          presets: [],
+        };
+
         if (pattern?.styles?.length) {
           const branchValues = pattern?.styles.map(s => {
             return {
@@ -94,27 +166,40 @@
               selected: !!(
                 branch &&
                 createBranchUrl(pattern.pattern, branch, s) === url &&
-                pattern.type === map.type
+                pattern.type === type
               ),
               pattern: pattern.pattern,
               dropdownId: hat(),
             };
           });
 
-          dropdownValues = dropdownValues.concat(branchValues);
-
-          dropdownDisplayOptions[
-            `Styles on a branch${pattern.name ? `: ${pattern.name}` : ''}`
-          ] = branchValues.map(item => ({
-            text: item.name,
-            dropdownId: item.dropdownId,
-          }));
+          preset.presets = preset.presets.concat(branchValues);
         }
+
+        patterns.push(preset);
+
+        dropdownValues = dropdownValues.concat(patterns);
+
+        dropdownDisplayOptions[`Styles on a branch`] = patterns.map(item => {
+          return {
+            text: item.name,
+            ...(item.type !== 'sublist' && { dropdownId: item.dropdownId }),
+            ...(item.type === 'sublist' && {
+              type: 'sublist',
+              presets: item.presets.map(v => ({
+                text: v.name,
+                dropdownId: v.dropdownId,
+              })),
+            }),
+          };
+        });
       }
     }
 
     // If no option has matched the URL, then it is custom
-    const hasSelectedOption = dropdownValues.some(item => !!item.selected);
+    const hasSelectedOption = dropdownValues.some(
+      item => !!item?.selected || item?.presets?.some(v => !!v.selected)
+    );
 
     // Create a custom value and display option
     const customValues = [
@@ -123,6 +208,7 @@
         dropdownType: 'custom',
         selected: !hasSelectedOption,
         type: 'mapbox-gl',
+        renderer: 'mapbox-gl',
         dropdownId: hat(),
       },
     ];
@@ -145,31 +231,74 @@
   stylePresetsStore.subscribe(value => {
     stylePresets = value;
     setInitialSelectedOption();
+    setRendererOptions();
   });
 
+  // On fetching a custom URL, update renderer options
+  const onSetUrl = e => {
+    selectedValue.url = e.detail.value;
+    setRendererOptions();
+  };
+
   const onSelectOption = e => {
+    // TODO needs to handle nesting
     const { dropdownId } = e.detail;
-    // Set selected property on value
-    dropdownValues = dropdownValues.map(v => ({
-      ...v,
-      selected: v.dropdownId === dropdownId,
-    }));
+
+    dropdownValues = dropdownValues.reduce((acc, v) => {
+      let next = v;
+      if (next?.type === 'sublist') {
+        next.presets = next.presets.map(v => ({
+          ...v,
+          selected: v.dropdownId === dropdownId,
+        }));
+      } else {
+        next = {
+          ...next,
+          selected: v.dropdownId === dropdownId,
+        };
+      }
+      acc.push(next);
+      return acc;
+    }, []);
 
     setSelectedValue();
+    setRendererOptions();
+  };
+
+  // Update map renderer in store
+  const updateMapRenderer = renderer => {
+    mapsStore.update(current => {
+      return current.map((m, i) => (i !== index ? m : { ...m, renderer }));
+    });
+  };
+
+  const onSelectRenderer = e => {
+    rendererValue = e.detail.value;
   };
 
   // Handle updating the map store
   const onUpdateMapStore = e => {
     const { value } = e.detail;
-    const nextMap = { ...value, index };
-    delete nextMap.dropdownId;
-    mapsStore.update(current => {
-      return current.map((m, i) => (i === index ? nextMap : m));
-    });
+    if (!value.isPolling) {
+      const nextMap = { ...value, index, renderer };
+      delete nextMap.dropdownId;
+      mapsStore.update(current => {
+        return current.map((m, i) => (i === index ? nextMap : m));
+      });
+    } else {
+      mapsStore.update(current => {
+        return current.map((m, i) => (i === index ? { ...m, ...value } : m));
+      });
+    }
   };
 
   const updateSelectedMapFromProps = nextValue => {
-    if (!nextValue) return;
+    const hasUnsetBranch =
+      selectedValue?.dropdownType === 'branch' && !selectedValue?.url;
+
+    // isPolling lets us continue polling while in the unset branch mode while letting
+    // us reset this on a history change as isPolling is never saved in the hash
+    if (!nextValue || (hasUnsetBranch && map.isPolling)) return;
     // Normally selectedValue should change map, but via props it's reversed
     // So that the bound selectedValue displays correctly
     if (
@@ -182,22 +311,29 @@
   };
 
   $: updateSelectedMapFromProps(map);
+
+  $: updateMapRenderer(rendererValue);
 </script>
 
-<div class="map-style-input">
+<div class="map-style-input-wrapper">
   {#if dropdownValues}
     <MapStyleInput
       dropdownValue={selectedValue}
       {dropdownDisplayOptions}
+      {rendererOptions}
+      {rendererValue}
+      {index}
       activeUrl={url}
+      on:setUrl={onSetUrl}
       on:selectOption={onSelectOption}
+      on:selectRenderer={onSelectRenderer}
       on:updateMapStore={onUpdateMapStore}
     />
   {/if}
 </div>
 
 <style>
-  .map-style-input {
+  .map-style-input-wrapper {
     margin-top: 6px;
   }
 </style>